Fixed caption edit bug
When navigating from one photo to the next while in edit mode, the previous photo's caption was displaying.

This was happening because Svelte doesn't create a new rich text editor component when going from one image to the next, but instead re-uses the existing editor, which contains the caption from the previous photo.